Added tkinter interface
